migration cleanups
