+++ backend/app/alembic/versions/057_v1_query_optimization.py
@@ -1,7 +1,7 @@
 """v1.0 query optimization: project_id + composite indexes, drop is_deleted
 
-Revision ID: 055
-Revises: 054
+Revision ID: 057
+Revises: 056
 Create Date: 2026-05-05 12:00:00.000000
 
 Bundles three coordinated changes for v1.0 lock:
@@ -34,8 +34,8 @@ import sqlalchemy as sa
 from alembic import op
 
 
-revision = "055"
-down_revision = "054"
+revision = "057"
+down_revision = "056"
 branch_labels = None
 depends_on = None
